Update TelemetryHandler.md
Changed HostOs and RuntimeEnvironment headers to be key/value pairs embedded in the comment.  New headers need to be on an allow-list and need to be accepted by CORS.  Embedding all the information in SdkVersion reduces the number of headers that need to be processed.  I do not believe any SDKs are currently emitting this telemetry